Restructured project as guided by tutorial. Added config and additional middleware
